Move use-database codegen into hooks subfolder, break out two util functions that will be used in upcoming domain hooks
