+++ src/scripts/codegen/constants.ts
@@ -1,4 +1,5 @@
 const BASE_IMPORT_PATH = "generated";
 const BASE_PATH = "src/generated";
+const TABLES_ENUM = "Tables";
 
-export { BASE_PATH, BASE_IMPORT_PATH };
+export { BASE_PATH, BASE_IMPORT_PATH, TABLES_ENUM };
