+++ src/scripts/codegen/utils.ts
@@ -3,6 +3,9 @@ import pluralize from "pluralize";
 import { PropertySignature } from "ts-morph";
 import { BASE_PATH } from "./constants";
 
+const getFromFunctionName = (property: PropertySignature): string =>
+    `from${getTableName(property)}`;
+
 const getInterfaceName = (property: PropertySignature): string =>
     _.capitalize(pluralize(property.getName(), 1));
 
@@ -12,4 +15,13 @@ const getInterfacePath = (property: PropertySignature): string =>
 const getInterfaceImportPath = (property: PropertySignature): string =>
     getInterfacePath(property).replace("src/", "").replace(".ts", "");
 
-export { getInterfaceName, getInterfacePath, getInterfaceImportPath };
+const getTableName = (property: PropertySignature): string =>
+    _.capitalize(property.getName());
+
+export {
+    getFromFunctionName,
+    getInterfaceImportPath,
+    getInterfaceName,
+    getInterfacePath,
+    getTableName,
+};
